\name{ais}
\alias{ais}
\docType{data}
\title{Australian institute of sport data}
\description{
Data on 102 male and 100 female athletes collected at the Australian
Institute of Sport.
}
\format{This data frame contains the following columns:
 \describe{
    \item{Sex}{
 (0 = male or 1 = female)
    }
    \item{Ht}{
 height (cm)
    }
    \item{Wt}{
 weight (kg)
    }
    \item{LBM}{
 lean body mass
    }
    \item{RCC}{
 red cell count
    }
    \item{WCC}{
 white cell count
    }
    \item{Hc}{
 Hematocrit
    }
    \item{Hg}{
 Hemoglobin
    }
    \item{Ferr}{
 plasma ferritin concentration
    }
    \item{BMI}{
 body mass index, weight/(height)**2
    }
    \item{SSF}{
 sum of skin folds
    }
    \item{Bfat}{
 Percent body fat
    }
    \item{Label}{
 Case Labels
    }
    \item{Sport}{
 Sport
    }
 }
}

\references{
S. Weisberg (2005).  \emph{Applied Linear Regression}, 3rd edition.  New York:
Wiley, Section 6.4}



\keyword{datasets}
